git monitor commit.

+++ nemu/include/cpu/instr/int.h
@@ -1,7 +1,7 @@
 #ifndef __INT_H__
 #define __INT_H__
 
-make_instr_func(int);
+make_instr_func(int_);
 
 
 #endif
